requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
faiss-cpu==1.9.0.post1
filelock==3.13.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.2.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pillow==10.2.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
PyPDF2==3.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
# scipy==1.14.1
sentence-transformers==3.3.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
voyageai==0.3.2
Werkzeug==3.1.3
yarl==1.18.3
